Dialogs/Settings: Split Tracking and add Services
SkyLines and LiveTrack24 are separate services, and together with
XCSoar Cloud and WeGlide they are identity and sharing settings, not
device setup. Keep xcsoar://config/tracking as a SkyLines alias.
Rename Setup Services to System Services so the labels do not collide.
Omit Quick Guide links when those features are not compiled in.
