+++ NEWS.txt
@@ -51,6 +51,9 @@ Version 7.46 - not yet released
   - fix the Devices Debug button not writing port traffic to xcsoar.log
     #2903
 * dialogs
+  - move SkyLines, LiveTrack24, XCSoar Cloud, and WeGlide to a new
+    Services configuration menu
+  - rename Setup Services to System Services
   - put the Close button on the right of the Status and Weather
     tab strips
   - label Close on the Quick Menu, BlueFly setup, and file Download
