+++ src/Dialogs/InternalLink.cpp
@@ -20,7 +20,16 @@
 #include "Dialogs/Settings/Panels/NetworkConfigPanel.hpp"
 #include "Dialogs/Settings/Panels/WeatherConfigPanel.hpp"
 #include "Dialogs/Settings/Panels/SafetyFactorsConfigPanel.hpp"
-#include "Dialogs/Settings/Panels/TrackingConfigPanel.hpp"
+#include "Tracking/Features.hpp"
+#ifdef HAVE_SKYLINES_TRACKING
+#include "Dialogs/Settings/Panels/SkyLinesConfigPanel.hpp"
+#endif
+#ifdef HAVE_LIVETRACK24
+#include "Dialogs/Settings/Panels/LiveTrack24ConfigPanel.hpp"
+#endif
+#ifdef HAVE_TRACKING
+#include "Dialogs/Settings/Panels/CloudConfigPanel.hpp"
+#endif
 #include "Dialogs/Settings/Panels/TerrainDisplayConfigPanel.hpp"
 #include "Dialogs/DataManagement/BackupRestorePanel.hpp"
 #include "Widget/Widget.hpp"
@@ -84,7 +93,19 @@ static constexpr ConfigPanelLink config_panel_links[] = {
   {"config/network",    N_("Network"),    CreateNetworkConfigPanel},
   {"config/weather",    N_("Weather"),    CreateWeatherConfigPanel},
   {"config/safety",     N_("Safety Factors"), CreateSafetyFactorsConfigPanel},
-  {"config/tracking",   N_("Tracking"),   CreateTrackingConfigPanel},
+#ifdef HAVE_SKYLINES_TRACKING
+  {"config/tracking",   "SkyLines",     CreateSkyLinesConfigPanel},
+  {"config/skylines",   "SkyLines",     CreateSkyLinesConfigPanel},
+#endif
+#ifdef HAVE_LIVETRACK24
+  {"config/livetrack24", "LiveTrack24", CreateLiveTrack24ConfigPanel},
+#ifndef HAVE_SKYLINES_TRACKING
+  {"config/tracking",   "LiveTrack24",  CreateLiveTrack24ConfigPanel},
+#endif
+#endif
+#ifdef HAVE_TRACKING
+  {"config/cloud",      N_("XCSoar Cloud"), CreateCloudConfigPanel},
+#endif
   {"config/terrain",    N_("Terrain Display"), CreateTerrainDisplayConfigPanel},
 };
