+++ src/Dialogs/Settings/Panels/LiveTrack24ConfigPanel.hpp
@@ -8,4 +8,4 @@
 class Widget;
 
 std::unique_ptr<Widget>
-CreateTrackingConfigPanel();
+CreateLiveTrack24ConfigPanel();
